Refactor and change of sorting to use name instead of title
